% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_read_agd.R
\name{batch_read_agd}
\alias{batch_read_agd}
\title{Reading several agd files and converting the settings and raw data tables to csv.}
\usage{
batch_read_agd(pathToDirectory, tz, allInOne)
}
\arguments{
\item{pathToDirectory}{path to the directory where all agd are saved.}

\item{tz}{Timezone. It must be the same for all files}

\item{allInOne}{Boolean indicating whether the output should be a list of list of two tables
(settings and raw data)
 or a list of two tables (settings and raw data for all files with a column filename)}
}
\description{
Reading several agd files and converting the settings and raw data tables to csv.
}
\details{
The function saves results in the working directory as csv files with a "," as separator.
If allInOne=TRUE, the output tables
 in this case will be called settings.csv and raw_data.csv. Else there will be a settings
 and a raw data file (originalfilename_settings.csv and originalfilename_raw.csv) for each
  file.
}
\examples{
\dontrun{
pathToDirectory <- system.file("extdata", package = "convertagd")
batch_read_agd(pathToDirectory, tz="GMT",
               allInOne=TRUE)
               }
}

