% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{read_agd}
\alias{read_agd}
\title{Reading a agd file}
\usage{
read_agd(file, tz = "GMT")
}
\arguments{
\item{file}{Full path to the agd file to be converted}

\item{tz}{Timezone of the measurements}
}
\value{
A list of two data tables with the settings and the measurements.
}
\description{
Reading a agd file
}
\examples{
file <- system.file("extdata", "dummyCHAI.agd", package = "convertagd")
testRes <- read_agd(file, tz = "GMT")
testRes[[1]]
testRes[[2]]
}

