% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_read_agd.R
\name{batch_read_agd}
\alias{batch_read_agd}
\title{Reading several agd files and converting the settings and raw data tables to csv.}
\usage{
batch_read_agd(path_to_directory, tz)
}
\arguments{
\item{path_to_directory}{path to the directory where all agd are saved.}

\item{tz}{Timezone. It must be the same for all files}
}
\description{
Reading several agd files and converting the settings and raw data tables to csv.
}
\details{
The function saves results in the input directory as csv files with a "," as separator.
Tthe output tables will be called settings.csv and raw_data.csv.
}
\examples{
\dontrun{
path_to_directory <- system.file("extdata", package = "convertagd")
batch_read_agd(path_to_directory, tz="GMT",
               all_in_one=TRUE)
               }
}

